[Xposed] Close .oat files after reading their header
Otherwise, resources will be exhausted sooner or later on some ROMs.
This can lead to bootloops or sudden restarts.

Some parts of the code were copied from OpenDexFilesFromImage(), which opens
the file for later use. But in OatHeader::FromFile(), we just want to read
the header and then close the file again.

Fixes rovo89/Xposed#31