[Xposed] Debugger fixes
Debuggers need to know about all methods in a class. For methods hooked with
Xposed, both the hooked (proxy-like) and the original method have to be returned.

Also, the hooked method isn't directly attached to a code item anymore, which
would otherwise be used to calculate offsets into the method code.

Fixes #1.