Parallel Move Resolver: Perform Stack/Stack first
On machines like x86, by the time other parallel moves are done, there
may be no free registers available to move/swap without having to save
and restore a register.

To avoid this, perform stack/stack first, while there is a good chance
that there is a destination register that we can use.  On the X86, this
avoids a lot of push eax/pop eax code.

Change-Id: I57076271b5672c931a93888ff23e30b2567f43b8
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>